+++ lib/src/analyzer_plugin/analyzer_plugin.dart
@@ -44,7 +44,7 @@ class MetricsAnalyzerPlugin extends ServerPlugin {
   String get name => 'Dart Code Metrics';
 
   @override
-  String get version => '1.5.1';
+  String get version => '1.6.0';
 
   @override
   void contentChanged(String path) {