+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-# Unreleased
+# 1.6.0
 - Add experimental static code diagnostics prefer-trailing-comma-for-collection, no-magic-number rule
 - Support number of arguments metric in analyzer plugin
 - Support excluding files from metrics calculation